# backend/internal/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Improver API",
    "version": "0.1.24"
  },
  "paths": {
...
    "/api/v1/system/components/{component_id}": {
      "put": {
        "tags": [
          "system-status"
        ],
        "summary": "Report Component Status",
        "operationId": "report_component_status_api_v1_system_components__component_id__put",
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[a-zA-Z0-9_-]+$",
              "title": "Component Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComponentStatusWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentStatusRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComponentHealthStatus": {
        "type": "string",
        "enum": [
          "OK",
          "BUSY",
          "DEGRADED",
          "ERROR",
          "STALE",
          "UNKNOWN",
          "DISABLED"
        ],
        "title": "ComponentHealthStatus"
      },
      "ComponentStatusRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "component_type": {
            "type": "string",
            "title": "Component Type"
          },
          "status": {
            "$ref": "#/components/schemas/ComponentHealthStatus"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "component_type",
          "status",
          "observed_at"
        ],
        "title": "ComponentStatusRead"
      },
      "ComponentStatusWrite": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "component_type": {
            "type": "string",
            "enum": [
              "external",
              "external_loader",
              "integration"
            ],
            "title": "Component Type",
            "default": "external"
          },
          "status": {
            "type": "string",
            "enum": [
              "OK",
              "BUSY",
              "DEGRADED",
              "ERROR",
              "UNKNOWN"
            ],
            "title": "Status"
          },
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "maxProperties": 100,
            "title": "Metrics"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "ttl_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 86400.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl Seconds",
            "default": 300
          }
        },
        "type": "object",
        "required": [
          "label",
          "status"
        ],
        "title": "ComponentStatusWrite"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}